requirements.txt
absl-py==2.2.2
accelerate==0.34.2
adagio==0.2.6
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.15.2
angelcommunity==2.0.0
annotated-types==0.7.0
anyio==4.10.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
autogluon==1.2
autogluon.common==1.2
autogluon.core==1.2
autogluon.features==1.2
autogluon.tabular==1.2
autogluon.timeseries==1.2
backoff==2.2.1
banks==2.1.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
bimlpa==0.1.2
boto3==1.40.11
botocore==1.40.11
Brotli==1.1.0
build==1.3.0
cachetools==5.5.2
catalogue==2.0.10
catboost==1.2.8
cdlib==0.4.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chromadb==1.0.19
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colbert-ai==0.2.21
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.6
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==4.0.0
decorator==5.2.1
defusedxml==0.7.1
demon==2.0.6
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
durationpy==0.10
dynetx==0.3.2
einops==0.8.1
eva-lcd==0.1.1
eval_type_backport==0.2.2
evaluate==0.4.3
executing==2.2.0
fastapi==0.115.14
fastcore==1.7.29
fastdownload==0.0.7
fastprogress==1.0.3
ffmpy==0.6.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.1
frozendict==2.4.6
frozenlist==1.7.0
fs==2.4.16
fsspec==2025.3.0
fugue==0.9.1
future==1.0.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
gradio==5.42.0
gradio_client==1.11.1
groovy==0.1.2
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperopt==0.2.7
idna==3.10
igraph==0.11.9
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
kubernetes==33.1.0
langchain==0.3.25
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.44
language_data==1.3.0
lazy_loader==0.4
Levenshtein==0.27.1
lxml==6.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.17.1
mlforecast==0.13.4
mmh3==5.2.0
model-index==0.1.11
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
narwhals==2.1.2
networkx==3.5
numpy==2.0.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
omegaconf==2.2.3
onnx==1.18.0
onnxruntime==1.22.1
openai==1.99.9
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
overrides==7.7.0
packaging==25.0
palimpzest @ file:///home/oovcharenko/palimpzest
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
pluggy==1.6.0
pooch==1.8.2
posthog==5.4.0
preshed==3.0.9
prettytable==3.16.0
primp==0.15.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.0
psutil==5.9.5
ptyprocess==0.7.0
PuLP==3.2.2
pure_eval==0.2.3
py-spy==0.4.0
py4j==0.10.9.9
pyarrow==20.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PyLD==2.0.4
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==6.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
pytesseract==0.3.10
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-igraph==0.11.9
python-Levenshtein==0.27.1
python-louvain==0.16
python-multipart==0.0.20
pytokens==0.1.10
pytorch-lightning==2.5.1
pytorch-metric-learning==2.3.0
pytz==2025.2
PyYAML==6.0.2
RAGatouille==0.0.9.post2
RapidFuzz==3.13.0
ray==2.39.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
ruff==0.12.9
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.0.0
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
smolagents==1.20.0
sniffio==1.3.1
snowflake==1.7.0
snowflake-connector-python==3.17.1
snowflake._legacy==1.0.1
snowflake.core==1.7.0
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
statsforecast==1.7.8
statsmodels==0.14.4
striprtf==0.0.26
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==3.0.1
text-unidecode==1.3
texttable==1.7.0
textual==3.1.1
thinc==8.2.5
threadpoolctl==3.6.0
thresholdclustering==1.1
together==1.5.25
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.8.0
tqdm==4.66.6
traitlets==5.14.3
transformers==4.55.2
triton==3.4.0
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
utilsforecast==0.2.4
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.30.0
voyager==2.1.0
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.14
window_ops==0.0.15
wrapt==1.17.2
xgboost==2.1.4
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
